% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrhenius.R
\name{arrhenius}
\alias{arrhenius}
\title{Fitting the Arrhenius temperature response model}
\usage{
arrhenius(param25, Ea, Tleaf)
}
\arguments{
\item{param25}{parameter value at 25 C}

\item{Ea}{activation energy in kJ/mol}

\item{Tleaf}{leaf temperature in K}
}
\value{
arrhenius is used to fit an exponential model
}
\description{
Fitting the Arrhenius temperature response model
}
