% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_topt_VJ.R
\name{fit_topt_VJ}
\alias{fit_topt_VJ}
\title{Fitting the temperature responses of Vcmax and Jmax}
\usage{
fit_topt_VJ(
  data,
  varnames = list(Vcmax = "Vcmax", Jmax = "Jmax", Tleaf = "Tleaf"),
  title = NULL,
  limit_jmax = 1e+05,
  limit_vcmax = 1e+05
)
}
\arguments{
\item{data}{Dataframe containing Vcmax, Jmax, and Tleaf}

\item{varnames}{Variable names}

\item{title}{Graph title, usually a group name}

\item{limit_jmax}{Upper limit to Jmax values for fitting}

\item{limit_vcmax}{Upper limit to Vcmax values for fitting}
}
\value{
fit_topt_VJ fits the Topt modified Arrhenius function to Vcmax and
Jmax data. Note that Hd may max out at 3000 kJ mol-1 for Jmax and 2000 kJ
mol-1 for Vcmax.

REFERENCE
Medlyn BE, Dreyer E, Ellsworth D, Forstreuter M, Harley PC,
Kirschbaum MUF, Le Roux X, Montpied P, Strassemeyer J, Walcroft A,
Wang K, Loutstau D. 2002. Temperature response of parameters of a
biochemically based model of photosynthesis. II. A review of
experimental data. Plant Cell Environ 25:1167-1179
}
\description{
Fitting the temperature responses of Vcmax and Jmax
}
\examples{
\dontrun{
Create Example Dataset

Tleaf <- seq(from = 10, to = 35, by = 5)
Tleaf <- sort(rep(Tleaf, 15))

Vcmax_start <- 90 * modarrhenius(Ea = 35, Hd = 220, dS = 0.3, Tleaf = Tleaf)
Jmax_start <- 140 * modarrhenius(Ea = 80, Hd = 220, dS = 0.6, Tleaf = Tleaf)
R_start <- 1.5 * arrhenius(Ea = 20, Tleaf = Tleaf)
GammaStar_start <- 42.75 * arrhenius(Ea = 37.83, Tleaf = Tleaf)
Km_start <- 718.40 * arrhenius(Ea = 65.50828, Tleaf = Tleaf)
Ci <- rep(c(75, 100, 150, 200, 250,
            300, 350, 400, 600, 800,
            1000, 1200, 1400, 1600, 1800), 6)
A <- 1:90
Vc <- 1:90
Vj <- 1:90
for(i in seq_along(A)){
  Vc[i] <- (Vcmax_start[i] * (Ci[i] - GammaStar_start[i]) / 
              (Ci[i] + Km_start[i]))
  Vj[i] <- (Jmax_start[i] * (Ci[i] - GammaStar_start[i]) / 
              (4 * Ci[i] + 8 * GammaStar_start[i]))
  if(Vc[i] < 0){
    Vc[i] <- NA
  }
  if(Vj[i] < 0){
    Vj[i] <- NA
  }
  A[i] <- min(Vc[i], Vj[i]) - R_start[i]
}

PPFD <- rep(2000, 90)
Press <- rep(100, 90)

data <- data.frame(cbind(A,
                         Ci,
                         Tleaf,
                         Press,
                         PPFD))

data$Treat <- as.character(Tleaf)


#Fit ACi Curves then fit temperature responses

fits <- fitacis2(data = data,
                 varnames = list(ALEAF = "A",
                                 Tleaf = "Tleaf",
                                 Ci = "Ci",
                                 PPFD = "PPFD",
                                 Rd = "Rd",
                                 Press = "Press"),
                 group1 = "Treat",
                 fitTPU = FALSE,
                 fitmethod = "bilinear",
                 gm25 = 10000,
                 Egm = 0)
outputs <- acisummary(data, group1 = "Treat", fits = fits)

tresp <- fit_topt_VJ(outputs)
}
}
