% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitacis2.R
\name{fitacis2}
\alias{fitacis2}
\title{Fit A-Ci curves with custom kinetics}
\usage{
fitacis2(data, group1, group2, group3, gm25, Egm, K25, Ek, fitmethod,
  fitTPU, Tcorrect, useRd = FALSE, citransition = NULL, alphag = 0,
  PPFD = NULL, Tleaf = NULL, alpha = 0.24, theta = 0.85,
  varnames = list(ALEAF = "Photo", Tleaf = "Tleaf", Ci = "Ci", PPFD =
  "PARi", Rd = "Rd"))
}
\arguments{
\item{data}{data frame with A/Ci curves}

\item{group1}{grouping variable 1, could be species, temperature, ID}

\item{group2}{grouping variable 2}

\item{group3}{grouping variable 3}

\item{gm25}{mesophyll conductance at 25C}

\item{Egm}{activation energy of mesophyll conductance in kJ/mol}

\item{K25}{Km at 25C}

\item{Ek}{activation energy of Km in kJ/mol}

\item{fitmethod}{See ?fitaci in plantecophys}

\item{fitTPU}{See ?fitaci in plantecophys. Set to TRUE/FALSE}

\item{Tcorrect}{See ?fitaci in plantecophys. Default here is FALSE}

\item{useRd}{See ?fitaci in plantecophys}

\item{citransition}{See ?fitaci in plantecophys}

\item{alphag}{See ?fitaci in plantecophys}

\item{PPFD}{See ?fitaci in plantecophys}

\item{Tleaf}{See ?fitaci in plantecophys}

\item{alpha}{See ?fitaci in plantecophys}

\item{theta}{See ?fitaci in plantecophys}

\item{varnames}{See ?fitaci in plantecophys}
}
\value{
fitacis2 allows gmeso, GammaStar, and Km to vary with Tleaf
        output matches the fitacis function from plantecophys
}
\description{
Fit A-Ci curves with custom kinetics
}
