% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_t_graphs.R
\name{get_t_graphs}
\alias{get_t_graphs}
\title{Get temperature response graphs}
\usage{
get_t_graphs(data)
}
\arguments{
\item{data}{List of data output from fit_topt_VJs}
}
\value{
get_t_pars returns temperature response graphs for
Vcmax and Jmax from the group fitting process
}
\description{
Get temperature response graphs
}
