% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toptfit.R
\name{toptfit}
\alias{toptfit}
\title{Fitting the Topt temperature response model}
\usage{
toptfit(Ea, Hd, kopt, Tleaf, Topt)
}
\arguments{
\item{Ea}{activation energy in kJ/mol}

\item{Hd}{deactivation energy in kJ/mol}

\item{kopt}{optimum parameter value}

\item{Tleaf}{leaf temperature in Celsius}

\item{Topt}{optimum leaf temperature in Celsius}

\item{param25}{parameter value at 25 C}
}
\value{
toptfit is used to fit a Topt model
}
\description{
Fitting the Topt temperature response model
}
