% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acisummary.R
\name{acisummary}
\alias{acisummary}
\title{Extracts coefficients from fitacis2}
\usage{
acisummary(data, group1, group2 = NA, group3 = NA, fits)
}
\arguments{
\item{data}{data frame with A/Ci curve data}

\item{group1}{grouping variable 1, must match fitacis2}

\item{group2}{grouping variable 2, must match fitacis2}

\item{group3}{grouping variable 3, must match fitacis2}

\item{fits}{list output from fitacis2}
}
\value{
acisummary produces a data frame with A-Ci coefficients
}
\description{
Extracts coefficients from fitacis2
}
\examples{
\dontrun{
data <- read.csv("mydata.csv")
fits <- fitacis2(data, group1 = "a", group2 = "b", group3 = "c",
fitmethod = "bilinear", fitTPU = TRUE, Tcorrect = FALSE)
outputs <- acisummary(data, group1 = "a", group2 = "b", group3 = "c", fits)
}

}
