% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrhenius.R
\name{arrhenius}
\alias{arrhenius}
\title{The Arrhenius temperature response equation}
\usage{
arrhenius(Ea, Tleaf)
}
\arguments{
\item{Ea}{activation energy in kJ/mol}

\item{Tleaf}{leaf temperature in Celsius}
}
\value{
arrhenius is an exponential model

Arrhenius S. 1915. Quantitative laws in biological chemistry. Bell.
}
\description{
The Arrhenius temperature response equation
}
