% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrhenius.R
\name{arrhenius}
\alias{arrhenius}
\title{Fitting the Arrhenius temperature response model}
\usage{
arrhenius(Ea, Tleaf)
}
\arguments{
\item{Ea}{activation energy in kJ/mol}

\item{Tleaf}{leaf temperature in Celsius}
}
\value{
arrhenius is used to fit an exponential model
}
\description{
Fitting the Arrhenius temperature response model
}
