% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_topt_VJ.R
\name{fit_topt_VJ}
\alias{fit_topt_VJ}
\title{Fitting the temperature responses of Vcmax and Jmax}
\usage{
fit_topt_VJ(
  data,
  varnames = list(Vcmax = "Vcmax", Jmax = "Jmax", Tleaf = "Tleaf"),
  title = NULL,
  limit_jmax = 1e+05,
  limit_vcmax = 1e+05
)
}
\arguments{
\item{data}{Dataframe containing Vcmax, Jmax, and Tleaf}

\item{varnames}{Variable names}

\item{title}{Graph title, usually a group name}

\item{limit_jmax}{Upper limit to Jmax values for fitting}

\item{limit_vcmax}{Upper limit to Vcmax values for fitting}
}
\value{
fit_topt_VJ fits the Topt modified Arrhenius function to Vcmax and
Jmax data. Note that Hd may max out at 3000 kJ mol-1 for Jmax and 2000 for
Vcmax.
}
\description{
Fitting the temperature responses of Vcmax and Jmax
}
