% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_topt_VJs.R
\name{fit_topt_VJs}
\alias{fit_topt_VJs}
\title{Fitting multiple temperature response curves}
\usage{
fit_topt_VJs(data, group, varnames = list(Vcmax = "Vcmax", Jmax = "Jmax",
  Tleaf = "Tleaf"), limit_jmax = 1e+05, limit_vcmax = 1e+05)
}
\arguments{
\item{data}{Dataframe with multiple temperature response curves}

\item{group}{Grouping variable to use}

\item{varnames}{Variable names}

\item{limit_jmax}{Upper limit to Jmax values for fitting}

\item{limit_vcmax}{Upper limit to Vcmax values for fitting}
}
\value{
fit_topt_VJs fits multiple Vcmax and Jmax temperature responses
}
\description{
Fitting multiple temperature response curves
}
