% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitacis2.R
\name{fitacis2}
\alias{fitacis2}
\title{Fit A-Ci curves with custom kinetics}
\usage{
fitacis2(
  data,
  group1,
  group2 = NA,
  group3 = NA,
  gm25 = 0.08701,
  Egm = 47.65,
  K25 = 718.4,
  Ek = 65.50828,
  Gstar25 = 42.75,
  Egamma = 37.83,
  fitmethod,
  fitTPU,
  Tcorrect,
  useRd = FALSE,
  citransition = NULL,
  alphag = 0,
  PPFD = NULL,
  Tleaf = NULL,
  alpha = 0.24,
  theta = 0.85,
  varnames = list(ALEAF = "Photo", Tleaf = "Tleaf", Ci = "Ci", PPFD = "PARi", Rd = "Rd")
)
}
\arguments{
\item{data}{data frame with A/Ci curves}

\item{group1}{grouping variable 1, could be species, temperature, ID}

\item{group2}{grouping variable 2}

\item{group3}{grouping variable 3}

\item{gm25}{mesophyll conductance at 25C}

\item{Egm}{activation energy of mesophyll conductance in kJ/mol}

\item{K25}{Km at 25C}

\item{Ek}{activation energy of Km in kJ/mol}

\item{Gstar25}{photorespiratory CO2 compensation point at 25 Celsius in ppm}

\item{Egamma}{activation energy of GammaStar in kJ/mol}

\item{fitmethod}{See ?fitaci in plantecophys}

\item{fitTPU}{See ?fitaci in plantecophys. Set to TRUE/FALSE}

\item{Tcorrect}{See ?fitaci in plantecophys. Default here is FALSE}

\item{useRd}{See ?fitaci in plantecophys}

\item{citransition}{See ?fitaci in plantecophys}

\item{alphag}{See ?fitaci in plantecophys}

\item{PPFD}{See ?fitaci in plantecophys}

\item{Tleaf}{See ?fitaci in plantecophys}

\item{alpha}{See ?fitaci in plantecophys}

\item{theta}{See ?fitaci in plantecophys}

\item{varnames}{See ?fitaci in plantecophys}
}
\value{
fitacis2 allows gmeso, GammaStar, and Km to vary with Tleaf
        output matches the fitacis function from plantecophys. Note
        that the temperature response function of Km is derived from
        the temperature responses of Ko and Kc in Bernacchi et al.
        2001, as is the GammaStar temperature response defaults. The
        gm defaults are from Bernacchi et al. 2002 fitted between 1
        and 35 Celsius. Also note that this ALWAYS uses gm. To fit
        data on a "Ci-basis", set gm25 really high (e.g. 10000) and
        Egm to 0.
}
\description{
Fit A-Ci curves with custom kinetics
}
\examples{
\dontrun{
data <- read.csv("mydata.csv")
fits <- fitacis2(data, group1 = "a", group2 = "b", group3 = "c",
fitmethod = "bilinear", fitTPU = TRUE, Tcorrect = FALSE)
}

}
