% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_t_pars.R
\name{get_t_pars}
\alias{get_t_pars}
\title{Get temperature response parameters}
\usage{
get_t_pars(data)
}
\arguments{
\item{data}{List of data output from fit_topt_VJs}
}
\value{
get_t_pars returns temperature response parameters for
Vcmax and Jmax from the group fitting process
}
\description{
Get temperature response parameters
}
