% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modarrhenius.R
\name{modarrhenius}
\alias{modarrhenius}
\title{Fitting the peaked Arrhenius temperature response model}
\usage{
modarrhenius(Ea, Hd, dS, Tleaf)
}
\arguments{
\item{Ea}{activation energy in kJ mol-1}

\item{Hd}{deactivation energy in kJ mol-1}

\item{dS}{entropy of deactivation in kJ mol-1}

\item{Tleaf}{leaf temperature in Celsius}
}
\value{
modarrhenius is used to fit a peaked Arrhenius model.
REFERENCE
Medlyn BE, Dreyer E, Ellsworth D, Forstreuter M, Harley PC,
Kirschbaum MUF, Le Roux X, Montpied P, Strassemeyer J, Walcroft A,
Wang K, Loutstau D. 2002. Temperature response of parameters of a
biochemically based model of photosynthesis. II. A review of
experimental data. Plant Cell Environ 25:1167-1179
}
\description{
Fitting the peaked Arrhenius temperature response model
}
