% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modarrhenius.R
\name{modarrhenius}
\alias{modarrhenius}
\title{Fitting the peaked Arrhenius temperature response model}
\usage{
modarrhenius(param25, Ea, Hd, dS, Tleaf)
}
\arguments{
\item{param25}{parameter value at 25 C}

\item{Ea}{activation energy in kJ/mol}

\item{Hd}{deactivation energy in kJ/mol}

\item{dS}{entropy parameter in kJ/mol}

\item{Tleaf}{leaf temperature in K}
}
\value{
modarrhenius is used to fit a peaked model
}
\description{
Fitting the peaked Arrhenius temperature response model
}
