% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_graphs.R
\name{print_graphs}
\alias{print_graphs}
\title{Printing graphs from a list of graphs}
\usage{
print_graphs(data)
}
\arguments{
\item{data}{List of graphs to output as .jpeg files}
}
\value{
print_graphs creates jpeg files from a list of graphs based on
the graph names
}
\description{
Printing graphs from a list of graphs
}
