% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toptfit.R
\name{toptfit}
\alias{toptfit}
\title{Fitting the Topt temperature response model}
\usage{
toptfit(Ea, Hd, kopt, Tleaf, Topt)
}
\arguments{
\item{Ea}{activation energy in kJ mol-1}

\item{Hd}{deactivation energy in kJ mol-1}

\item{kopt}{parameter value at optimum temperature}

\item{Tleaf}{leaf temperature in Celsius}

\item{Topt}{optimum leaf temperature in Celsius}
}
\value{
toptfit is the Topt temperature response model.
REFERENCE
Medlyn BE, Dreyer E, Ellsworth D, Forstreuter M, Harley PC,
Kirschbaum MUF, Le Roux X, Montpied P, Strassemeyer J, Walcroft A,
Wang K, Loutstau D. 2002. Temperature response of parameters of a
biochemically based model of photosynthesis. II. A review of
experimental data. Plant Cell Environ 25:1167-1179
}
\description{
Fitting the Topt temperature response model
}
